updated volumes for newton
Took the volumes code we had which was from
mitaka and did a diff of that with a default
mitaka install and came up with this patchset:
https://gist.github.com/michaelrice/28d5f8af7d3a33b8de94405c27183350
Next I applied this patch set to a default Newton
and fixed the few merge conflicts. Next copied
the volumes code back into this project.